refactor: Rename pending-submissions's exports file
